refactor: vote -> hotVote 변경 및 props 구조 변경
