+++ src/components/Home/HotVoteList/HotVoteListItem.styles.tsx
